Rename keyboards filenames
